-hist command implemented (#40)
* Added -hist command, which shows the user their history stack in fzf allowing them to navigate to a selected directory in their history
* Added docker containers for test environments, set up manual testing for bash function
* Added makefile to easily run above mentioned test environments
* Adjusted bash functions to allow for easier testing
* Added bash function to alert the user fzf is a dependency for using the -hist command
* Changed cd to pushd to account for shells where cd does not automatically push to dirs stack
* Adjusted changeDirectory to not care about length of allPaths map since it handles more than just key evaluation
* Adjusted binary file to be lazy loaded for commands that read/write to it

+++ shells/bash/ftmain.sh
@@ -1,41 +1,63 @@
 #!/usr/bin/env bash
 
-upperStack=()
+ft__check_fzf() {
+    if ! command -v fzf &>/dev/null; then
+        echo "Error: fzf is needed to use this feature. You can install it using your package manager:"
+        echo "  - Ubuntu/Debian: sudo apt install fzf"
+        echo "  - Fedora: sudo dnf install fzf"
+        echo "  - Arch: sudo pacman -S fzf"
+        echo "  - macOS: brew install fzf"
+        return 1
+    fi
+}
 
-pushup() {
+ft__upperStack=()
+
+ft__pushup() {
     local _path="$1"
-    upperStack+=("$_path")
+    ft__upperStack+=("$_path")
 }
 
-popup() {
-    unset 'upperStack[-1]'
-    upperStack=(${upperStack[@]})
+ft__popup() {
+    unset 'ft__upperStack[-1]'
+    ft__upperStack=(${ft_upperStack[@]})
 }
 
-ft() {
-
-    temp_output=$(mktemp)
+ft__capture() {
+    local temp_output=$(mktemp)
     
     "$FT_EXE_PATH" "$@" | tee "$temp_output"
 
-    output="$(tail -n 1 "$temp_output")"
+    local output="$(tail -n 1 "$temp_output")"
+    echo "$output" "$temp_output"
+}
+
+ft__execute() {
+    local output="$1"
+    local temp_output="$2"
 
     if [[ -d "$output" || "$output" == ".." || "$output" == "-" ]]; then 
-        upperStack=()
-        cd "$output";
+        
+        ft__upperStack=()
+        pushd "$output" > /dev/null
+
     elif [[ "$output" == "]" ]]; then    
-        if [ ${#upperStack[@]} -eq 0 ]; then
+        
+        if [ ${#ft__upperStack[@]} -eq 0 ]; then
             echo Already at head of history stack.
             rm "$temp_output"
             return 1
         
         fi
-        local p="${upperStack[-1]}"
-        popup
-        cd "$p"
+
+        local p="${ft__upperStack[-1]}"
+        ft__popup
+        pushd "$p" > /dev/null
 
     elif [[ "$output" == "[" ]]; then
+        
         local lowerStackLen=$(dirs -v | awk '{print $1}' | sort -n | tail -1)
+        
         if [ "$lowerStackLen" -eq 0 ]; then
             echo Already at tail of history stack.
             rm "$temp_output"
@@ -43,14 +65,40 @@ ft() {
         fi
 
         local p=$(pwd)
-        pushup "${p}"
+        ft__pushup "${p}"
         popd > /dev/null
     
-    fi
+    elif [[ "$output" == "-hist" ]]; then
+        # TODO: test me!!
+        ft__check_fzf
+        
+        local navigation="up;"
+        for i in "${ft__upperStack[@]}"; do
+            navigation="$navigation;"
+            navigation="$navigation down"
+        done
 
-    rm "$temp_output"
+        local lowerStack=($(dirs -v | awk '{print $2}'))
+        local historyStack=("${ft__upperStack[@]}" "${lowerStack[@]}")
+
+        local selected=$(printf "%s\n" "${historyStack[@]}" | fzf \
+            --bind "start:execute-silent($navigation)")
+        
+        if [[ -n "$selected" ]]; then
+            pushd "$selected" > /dev/null || echo "Could not find directory $selected"
+        fi
+
+    fi
+    
+    if [[ -f "$temp_output" ]]; then
+        rm "$temp_output"
+    fi
 
 }
 
 
+ft() {
+    local cmd=$(ft__capture)
+    ft__execute "$cmd"
+}
